\name{LPGraph-package}
\alias{LPGraph-package}
\alias{LPGraph}
\docType{package}
\title{
\packageTitle{LPGraph}
}
\description{
    This package provides an algorithm to approximate and smooth the Laplacian graph spectra of a network with ordered vertices using LP-nonparametric basis. The underlying theory yields a reduced order (compressive) characterization of spectral graph properties.
}
\author{
Mukhopadhyay, S. and Wang, K.

Maintainer: Kaijun Wang <kaijun.wang@temple.edu>
}
\references{
Mukhopadhyay, S. and Wang, K. (2018), "Graph Spectral Compression via Smoothing".

Mukhopadhyay, S. (2017+), "Unified Statistical Theory of Spectral Graph Analysis".

Mukhopadhyay, S. and Parzen, E. (2014), "LP Approach to Statistical Modeling", arXiv:1405.2601.
}

\keyword{ package }
